Fix: Rename Markdown-related Files and Imports for Consistency and Clarity (#8767)
# Description 

This PR renames the Markdown file for better consistency of codebase and
to avoid confusion while making imports.

Fixes: #8237 

# Comments

 ### **Changes made** 
- Changed file names
- _Markdown.vue -> ProblemMarkdown.vue_ in
`frontend/www/js/omegaup/components/problem/Markdown.vue`
  
- Updated imports in various files
   - `frontend/www/js/omegaup/components/course/ViewStudent.vue` 
   - `frontend/www/js/omegaup/components/problem/Detailsv2.vue`
   - `frontend/www/js/omegaup/components/problem/Details.vue`
   - `frontend/www/js/omegaup/components/problem/Edit.vue` 
   - `frontend/www/js/omegaup/components/problem/Markdown.test.ts` 
   - `frontend/www/js/omegaup/components/problem/Print.vue` 
   - `frontend/www/js/omegaup/components/problem/Solution.vue` 
   - `frontend/www/js/omegaup/components/problem/StatementEdit.vue` 
-
`frontend/www/js/omegaup/components/problem/creator/solution/SolutionTab.vue`
-
`frontend/www/js/omegaup/components/problem/creator/statement/StatementTab.vue`

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/blob/main/frontend/www/docs/Coding-guidelines.md)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.


# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/blob/main/frontend/www/docs/Coding-guidelines.md)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.
