+++ frontend/www/js/omegaup/components/problem/ProblemMarkdown.vue
@@ -74,7 +74,6 @@ export default class ProblemMarkdown extends Vue {
 
   @Emit('rendered')
   private renderMathJax(): void {
-    // console.log('[Markdown.vue] Rendered HTML:', this.html);
     this.root.innerHTML = this.html;
     this.root
       .querySelectorAll(
